refactor(autocomplete): move autocomplete test helpers into separate file
